Prepare for SIMD
