+++ src/tools/bench.rs
@@ -78,5 +78,18 @@ pub fn bench(td: &mut ThreadData) {
     td.debug_stats.print();
     td.clear();
     let nps = (nodes / time) * 1000;
+
+    #[cfg(target_feature = "avx512f")]
+    println!("AVX512F");
+
+    #[cfg(all(target_feature = "avx2", not(target_feature = "avx512f")))]
+    println!("AVX2");
+
+    #[cfg(target_feature = "neon")]
+    println!("NEON");
+
+    #[cfg(not(any(target_feature = "avx512f", target_feature = "avx2", target_feature = "neon")))]
+    println!("Scalar");
+
     println!("{} nodes {} nps", nodes, nps);
 }
